+++ src/primaite/simulator/network/hardware/nodes/computer.py
@@ -1,4 +1,6 @@
 from primaite.simulator.network.hardware.base import NIC, Node
+from primaite.simulator.system.services.dns.dns_client import DNSClient
+from primaite.simulator.system.services.ftp.ftp_client import FTPClient
 
 
 class Computer(Node):
@@ -36,3 +38,14 @@ class Computer(Node):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         self.connect_nic(NIC(ip_address=kwargs["ip_address"], subnet_mask=kwargs["subnet_mask"]))
+        self._install_system_software()
+
+    def _install_system_software(self):
+        """Install System Software - software that is usually provided with the OS."""
+        # DNS Client
+        self.software_manager.install(DNSClient)
+
+        # FTP
+        self.software_manager.install(FTPClient)
+
+        super()._install_system_software()